Assume install_dir is $HOME
In the future, any dir should be supported as home dir. The problem
is that we then need to know, after installation, which one that is
which isn't as obvious as it might seem. This is because syscond is
commonly started via a symbolic link in ~/bin, which gives syscond
the impression that it's running from that directory.

This is probably best solved with a ~/.syscondrc file, or perhaps
it should be placed in a more modern file location.